+++ CHANGES.md
@@ -2,7 +2,7 @@ Changes in Element 1.0.9 (2020-XX-XX)
 ===================================================
 
 Features ✨:
- -
+ - Support for the Conversations feature in Android 11 (#1722)
 
 Improvements 🙌:
  -